Update successful job logic; address #122, #128, #129.
- Updates conditional to Spark job for success
- Adds failed job mailer for a failed Spark job
- Updates delayed_job config (let a job fail once)
- Add a bunch of conditional logic on collection page to display what we
have, and ignore what we don't have
- Add a bunch of helpers that allow for the conditions in the
collections view (needs lots of DRY!)
- Update Rubocop for this chaos

+++ app/jobs/collections_cat_job.rb
@@ -14,7 +14,7 @@ class CollectionsCatJob < ApplicationJob
       logger.info 'Executing: ' + combine_full_url_output_cmd
       system(combine_full_url_output_cmd)
       successful_job = collection_derivatives + '/all-domains/' + c.collection_id.to_s + '-fullurls.txt'
-      if File.exist? successful_job
+      if File.exist?(successful_job) && !File.empty?(successful_job)
         FileUtils.rm_rf(collection_derivatives + '/all-domains/output')
       else
         raise 'Collections cat domains job failed.'
@@ -22,8 +22,7 @@ class CollectionsCatJob < ApplicationJob
       combine_full_text_output_cmd = 'cat ' + collection_derivatives + '/all-text/output/part* > ' + collection_derivatives + '/all-text/' + c.collection_id.to_s + '-fulltext.txt'
       logger.info 'Executing: ' + combine_full_text_output_cmd
       system(combine_full_text_output_cmd)
-      successful_job = collection_derivatives + '/all-text/' + c.collection_id.to_s + '-fulltext.txt'
-      if File.exist? successful_job
+      if File.exist?(successful_job) && !File.empty?(successful_job)
         FileUtils.rm_rf(collection_derivatives + '/all-text/output')
       else
         raise 'Collections cat full text job failed.'
